Refactor,Extract Common Functionality into BaseCausalLM Abstract Class
